Add Kafka configuration, payment processing, and booking event handling

+++ booking-service/src/main/java/serg/madi/bookingservice/dto/request/BookingRequest.java
@@ -15,5 +15,5 @@ public class BookingRequest {
     LocalDate checkIn;
     LocalDate checkOut;
     Integer peopleCount;
-    Float price;
+    Double price;
 }
